cpp: test: Split units tests into separate files
The template expansions can exceed the 16-bit limit on FreeBSD and
Windows, particularly when we instantiate and interconvert all unit
types.
